requirement.txt
asgiref==3.5.2
cached-property==1.5.2
colorama==0.4.6
Django==4.1
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-dynamic-fields==0.4.0
enum-compat==0.0.3
importlib-metadata==5.1.0
loguru==0.6.0
Markdown==3.4.1
PyJWT==1.7.1
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
win32-setctime==1.1.0
zipp==3.11.0